Read Configuration File

Reads the values in the configuration file created by batch_config().

batch_config_read(path)

Arguments

path
A string of the path to the directory with the files for processing.

Value

A named list of the configuration values.

See also

Examples

path <- tempdir()
write.csv(mtcars, file.path(path, "file1.csv"))
batch_config(function(x) TRUE, path, regexp = "[.]csv$")
batch_config_read(path)
#> $time
#> [1] "2021-09-20 15:04:38 UTC"
#> 
#> $regexp
#> [1] "[.]csv$"
#> 
#> $recurse
#> [1] FALSE
#> 
#> $fun
#> function(x) TRUE
#> <environment: 0x7fe584313e40>
#> 
#> $dots
#> list()
#> 
batch_cleanup(path, force = TRUE, remaining = TRUE)
unlink(file.path(path, "file1.csv"))