Batch File Text Replacement

Uses batch_process() and gsub_file() to perform batch text file replacement. By default it replaces text in all .R and .r files in the working directory.

batch_gsub(pattern, replacement, path = ".", regexp = "[.](R|r)$",
  recurse = FALSE, progress = FALSE,
  options = furrr::future_options(), ask = getOption("batchr.ask",
  TRUE))

Arguments

pattern
A string of the regular expression to match.

replacement
A string of the replacement text.

path
A string of the path to the directory with the files for processing.

regexp
A string of a regular expression. Only non-hidden file names which match the regular expression will be batch processed.

recurse
A flag specifying whether to recurse into path's subdirectories.

progress
A flag specifying whether to print a progress bar.

options
The future specific options to use with the workers. seed must be FALSE.

ask
A flag specifying whether to ask before starting to process the files.

Value

An invisible flag indicating whether all the files were successfully processed.

See also